ENGR00217318-3 flexcan: only enter stop mode when device is up
The flexcan is still in disable mode during suspend if it's still
not up. We do not need to enter stop mode if find the device is not
up since the stop mode does not work well in disable mode(remote wakeup
does not work).

Using disable mode for suspend if it's not up.

Signed-off-by: Dong Aisheng <b29396@freescale.com>